payment/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0